chore: move new function to IR_tools
